requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
